Pretrain/requirements/mminstall.txt
mmcv-full>=1.3.8,<=1.4.0
mmdet>=2.14.0,<=3.0.0
